Uniquify candidates
